// frontend/src/api/models/MainProfileResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Level-Up API
 * API for Level-Up app
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MainProfileResponse
 */
export interface MainProfileResponse {
    /**
     * 
     * @type {string}
     * @memberof MainProfileResponse
     */
    dateOfBirth?: string;
    /**
     * 
     * @type {string}
     * @memberof MainProfileResponse
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof MainProfileResponse
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof MainProfileResponse
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof MainProfileResponse
     */
    nickName: string;
}

/**
 * Check if a given object implements the MainProfileResponse interface.
 */
export function instanceOfMainProfileResponse(value: object): value is MainProfileResponse {
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('firstName' in value) || value['firstName'] === undefined) return false;
    if (!('lastName' in value) || value['lastName'] === undefined) return false;
    if (!('nickName' in value) || value['nickName'] === undefined) return false;
    return true;
}

export function MainProfileResponseFromJSON(json: any): MainProfileResponse {
    return MainProfileResponseFromJSONTyped(json, false);
}

export function MainProfileResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MainProfileResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'dateOfBirth': json['dateOfBirth'] == null ? undefined : json['dateOfBirth'],
        'email': json['email'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'nickName': json['nickName'],
    };
}

export function MainProfileResponseToJSON(json: any): MainProfileResponse {
    return MainProfileResponseToJSONTyped(json, false);
}

export function MainProfileResponseToJSONTyped(value?: MainProfileResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dateOfBirth': value['dateOfBirth'],
        'email': value['email'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'nickName': value['nickName'],
    };
}